works without a double type variable now

+++ 0x03-debugging/2-largest_number.c
@@ -8,9 +8,9 @@
  * Return: largest number
  */
 
-double largest_number(double a, double b, double c)
+int largest_number(int a, int b, int c)
 {
-	double largest;
+	int largest;
 
 	if (c >= a && c >= b)
 	{
